requirements.txt
aniso8601==7.0.0
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-graphql-playground==1.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==2.10
promise==2.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
Rx==1.6.1
simplejson==3.17.3
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
